Remove WHyred. So long nerds
Sorry users! I have no time and my device display is broken. I am not willing to do blind builds and so i am dropping this for now Maybee in future if i fix it or if someone supports me then i will support it again. Dropping this so that if other devs are free they can do better job